fix(sequelize): :bug: sales and salesProduct field type changed
the price and total price fields were integers and had to be decimal numbers

+++ back-end/src/database/models/products.js
@@ -11,7 +11,7 @@ const ProductModel = (sequelize, DataTypes) => {
       allowNull: false,
     },
     price: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(4,2),
       allowNull: false,
     },
     urlImage: {
